Extract CollectibleGatedAvailability
